# openapi.json
{
  "components": {
    "schemas": {
      "ActionRequest": {
        "properties": {
          "kind": {
            "enum": [
              "LEAVE",
              "HIDE",
              "UNHIDE",
              "DELETE"
            ],
            "title": "Kind",
            "type": "string"
          }
        },
        "required": [
          "kind"
        ],
        "title": "ActionRequest",
        "type": "object"
      },
...
      "CorrectionRequest": {
        "properties": {
          "note": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Note"
          },
          "severity": {
            "enum": [
              "SAFE",
              "OFFENSIVE",
              "HARMFUL"
            ],
            "title": "Severity",
            "type": "string"
          },
          "target": {
            "enum": [
              "PERSON",
              "INSTITUTION",
              "NEITHER"
            ],
            "title": "Target",
            "type": "string"
          }
        },
        "required": [
          "severity",
          "target"
        ],
        "title": "CorrectionRequest",
        "type": "object"
      },
      "CorrectionResponse": {
        "properties": {
          "actor": {
            "title": "Actor",
            "type": "string"
          },
          "disagrees_with_model": {
            "title": "Disagrees With Model",
            "type": "string"
          },
          "note": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Note"
          },
          "occurred_at": {
            "format": "date-time",
            "title": "Occurred At",
            "type": "string"
          },
          "severity": {
            "title": "Severity",
            "type": "string"
          },
          "target": {
            "title": "Target",
            "type": "string"
          }
        },
        "required": [
          "severity",
          "target",
          "actor",
          "note",
          "disagrees_with_model",
          "occurred_at"
        ],
        "title": "CorrectionResponse",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "HistoryEntry": {
        "properties": {
          "actor": {
            "title": "Actor",
            "type": "string"
          },
          "kind": {
            "title": "Kind",
            "type": "string"
          },
          "occurred_at": {
            "format": "date-time",
            "title": "Occurred At",
            "type": "string"
          }
        },
        "required": [
          "kind",
          "actor",
          "occurred_at"
        ],
        "title": "HistoryEntry",
        "type": "object"
      },
...
      "ReasonCount": {
        "properties": {
          "count": {
            "title": "Count",
            "type": "integer"
          },
          "surfaced_reason": {
            "title": "Surfaced Reason",
            "type": "string"
          }
        },
        "required": [
          "surfaced_reason",
          "count"
        ],
        "title": "ReasonCount",
        "type": "object"
      },
...
      "SampleRemoval": {
        "properties": {
          "removed": {
            "title": "Removed",
            "type": "integer"
          }
        },
        "required": [
          "removed"
        ],
        "title": "SampleRemoval",
        "type": "object"
      },
...
      "Summary": {
        "description": "Counts computed across the whole workspace, not from one page.",
        "properties": {
          "deleted": {
            "title": "Deleted",
            "type": "integer"
          },
          "left_visible": {
            "title": "Left Visible",
            "type": "integer"
          },
          "need_review": {
            "title": "Need Review",
            "type": "integer"
          },
          "pending": {
            "title": "Pending",
            "type": "integer"
          },
          "processed": {
            "title": "Processed",
            "type": "integer"
          },
          "reasons": {
            "items": {
              "$ref": "#/components/schemas/ReasonCount"
            },
            "title": "Reasons",
            "type": "array"
          },
          "reviewed": {
            "title": "Reviewed",
            "type": "integer"
          }
        },
        "required": [
          "processed",
          "need_review",
          "reviewed",
          "pending",
          "left_visible",
          "deleted",
          "reasons"
        ],
        "title": "Summary",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
      "WorkList": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/WorkListItem"
            },
            "title": "Items",
            "type": "array"
          },
          "limit": {
            "title": "Limit",
            "type": "integer"
          },
          "offset": {
            "title": "Offset",
            "type": "integer"
          },
          "total": {
            "title": "Total",
            "type": "integer"
          }
        },
        "required": [
          "items",
          "total",
          "limit",
          "offset"
        ],
        "title": "WorkList",
        "type": "object"
      },
...
    }
  },
  "info": {
    "title": "KCMS Backend",
    "version": "1.0.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/comments": {
      "get": {
        "operationId": "listComments",
        "parameters": [
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 25,
              "maximum": 100,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "offset",
            "required": false,
            "schema": {
              "default": 0,
              "minimum": 0,
              "title": "Offset",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "maxLength": 200,
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Query"
            }
          },
          {
            "in": "query",
            "name": "severity",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "enum": [
                    "SAFE",
                    "OFFENSIVE",
                    "HARMFUL"
                  ],
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Severity"
            }
          },
          {
            "in": "query",
            "name": "target",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "enum": [
                    "PERSON",
                    "INSTITUTION",
                    "NEITHER"
                  ],
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Target"
            }
          },
          {
            "in": "query",
            "name": "surfaced_reason",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "enum": [
                    "triage",
                    "institution_sample",
                    "novel_language",
                    "uncertainty"
                  ],
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Surfaced Reason"
            }
          },
          {
            "in": "query",
            "name": "review_status",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "enum": [
                    "PENDING",
                    "ACTIONED"
                  ],
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Review Status"
            }
          },
          {
            "in": "query",
            "name": "sort",
            "required": false,
            "schema": {
              "default": "PRIORITY",
              "enum": [
                "PRIORITY",
                "NEWEST",
                "OLDEST"
              ],
              "title": "Sort",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WorkList"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Comments"
      }
    },
    "/api/v1/comments/samples": {
      "delete": {
        "description": "Clear the seeded sample comments from this workspace.\n\nOnly the samples are removed: the delete is scoped by the sample Page id,\nso comments imported from a connected Facebook Page are never touched.\nEmptying a shared workspace affects everyone in it, so it is an owner\naction rather than something any member can do.",
        "operationId": "removeSampleComments",
        "parameters": [
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SampleRemoval"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Remove Sample Comments"
      }
    },
    "/api/v1/comments/summary": {
      "get": {
        "operationId": "getWorkspaceSummary",
        "parameters": [
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Summary"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Summary"
      }
    },
    "/api/v1/comments/{comment_id}/actions": {
      "post": {
        "description": "Records a moderation Action. Actions are append-only and reversible,\nand never become training labels.\n\nWhen the comment came from a connected Facebook Page, HIDE, UNHIDE and\nDELETE are applied on Facebook as well. The Action row and the Facebook\nstate are written together: if Facebook refuses, the row is rolled back,\nbecause an Action records what actually happened to the comment.\n\nHIDE is reversible through UNHIDE and is the option to reach for when the\nclassifier's judgement is uncertain. DELETE cannot be undone on Facebook.\nLEAVE changes nothing there and exists so that a decision to allow a\ncomment is still recorded, which is a different fact from nobody having\nlooked.\n\nAny action here also cancels a pending auto-delete on this comment\n(quarantine.sweep_once would otherwise delete it once the delay expires):\na human has now decided, so the countdown no longer applies.",
        "operationId": "recordAction",
        "parameters": [
          {
            "in": "path",
            "name": "comment_id",
            "required": true,
            "schema": {
              "title": "Comment Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ActionRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/HistoryEntry"
                  },
                  "title": "Response Recordaction",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Record Action"
      }
    },
    "/api/v1/comments/{comment_id}/corrections": {
      "post": {
        "description": "Records what a human asserts the labels should be.\n\nA Correction is not an Action. Submitting one does not hide, unhide or\notherwise change what happens to the comment.",
        "operationId": "recordCorrection",
        "parameters": [
          {
            "in": "path",
            "name": "comment_id",
            "required": true,
            "schema": {
              "title": "Comment Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CorrectionRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CorrectionResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Record Correction"
      }
    },
...
  }
}